Move unit tests from QuestionDefinitionTest to new subclasses
